revert: remove shared layout
